refactor: move sync implementation into core src
