+++ packages/core/src/sync/providers/google.ts
@@ -1,6 +1,6 @@
 import { z } from "zod";
 
-import type { ExistingModel, SyncProvider, SyncedModel } from "../sync-models.js";
+import type { ExistingModel, SyncProvider, SyncedModel } from "../index.js";
 
 const API_ENDPOINT = "https://generativelanguage.googleapis.com/v1beta/models";
